exgentic_a2a_runner/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "exgentic-a2a-runner"
version = "0.1.0"
description = "Exgentic Benchmark A2A Runner for Kagenti"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.9.0",
    "a2a-sdk>=0.3.0,<1.0",
    "httpx>=0.27.0",
    "requests>=2.28.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-requests>=0.41b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["exgentic_a2a_runner*"]

[project.scripts]
exgentic-a2a-runner = "exgentic_a2a_runner.runner:main"